requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
psycopg2==2.9.5
pycodestyle==2.10.0
pyngrok==5.2.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uWSGI>=2.0.18,<2.1